+++ cel-sampler/src/main/java/io/opentelemetry/contrib/sampler/cel/internal/CelBasedSamplerComponentProvider.java
@@ -10,8 +10,8 @@ import io.opentelemetry.api.incubator.config.DeclarativeConfigException;
 import io.opentelemetry.api.incubator.config.DeclarativeConfigProperties;
 import io.opentelemetry.contrib.sampler.cel.CelBasedSampler;
 import io.opentelemetry.contrib.sampler.cel.CelBasedSamplerBuilder;
+import io.opentelemetry.sdk.autoconfigure.declarativeconfig.DeclarativeConfiguration;
 import io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider;
-import io.opentelemetry.sdk.extension.incubator.fileconfig.DeclarativeConfiguration;
 import io.opentelemetry.sdk.trace.samplers.Sampler;
 import java.util.List;