+++ ibm-mq-metrics/src/test/java/io/opentelemetry/ibm/mq/metricscollector/QueueCollectionBuddyTest.java
@@ -15,32 +15,29 @@ import com.ibm.mq.constants.CMQC;
 import com.ibm.mq.constants.CMQCFC;
 import com.ibm.mq.headers.pcf.PCFMessage;
 import com.ibm.mq.headers.pcf.PCFMessageAgent;
-import io.opentelemetry.api.metrics.Meter;
 import io.opentelemetry.ibm.mq.config.QueueManager;
+import io.opentelemetry.ibm.mq.metrics.MetricProducer;
 import io.opentelemetry.ibm.mq.metrics.MetricsConfig;
 import io.opentelemetry.ibm.mq.opentelemetry.ConfigWrapper;
+import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
 import io.opentelemetry.sdk.metrics.data.LongPointData;
 import io.opentelemetry.sdk.metrics.data.MetricData;
-import io.opentelemetry.sdk.testing.junit5.OpenTelemetryExtension;
+import io.opentelemetry.sdk.resources.Resource;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.api.extension.RegisterExtension;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class QueueCollectionBuddyTest {
-  @RegisterExtension
-  static final OpenTelemetryExtension otelTesting = OpenTelemetryExtension.create();
 
   QueueCollectionBuddy classUnderTest;
   QueueManager queueManager;
   MetricsCollectorContext collectorContext;
-  Meter meter;
   @Mock private PCFMessageAgent pcfMessageAgent;
 
   @BeforeEach
@@ -48,7 +45,6 @@ class QueueCollectionBuddyTest {
     ConfigWrapper config = ConfigWrapper.parse("src/test/resources/conf/config.yml");
     ObjectMapper mapper = new ObjectMapper();
     queueManager = mapper.convertValue(config.getQueueManagers().get(0), QueueManager.class);
-    meter = otelTesting.getOpenTelemetry().getMeter("opentelemetry.io/mq");
     collectorContext =
         new MetricsCollectorContext(queueManager, pcfMessageAgent, null, new MetricsConfig(config));
   }
@@ -61,8 +57,9 @@ class QueueCollectionBuddyTest {
     sharedState.putQueueType("DEV.QUEUE.1", "local-transmission");
     PCFMessage request = createPCFRequestForInquireQStatusCmd();
     when(pcfMessageAgent.send(request)).thenReturn(createPCFResponseForInquireQStatusCmd());
-
-    classUnderTest = new QueueCollectionBuddy(meter, sharedState);
+    MetricProducer producer =
+        new MetricProducer(Resource.empty(), InstrumentationScopeInfo.empty());
+    classUnderTest = new QueueCollectionBuddy(producer, sharedState);
     classUnderTest.processPcfRequestAndPublishQMetrics(
         collectorContext, request, "*", InquireQStatusCmdCollector.ATTRIBUTES);
 
@@ -72,6 +69,7 @@ class QueueCollectionBuddyTest {
                 "DEV.DEAD.LETTER.QUEUE",
                 new HashMap<>(
                     ImmutableMap.of(
+                        "ibm.mq.expired.messages", 2L,
                         "ibm.mq.oldest.msg.age", -1L,
                         "ibm.mq.uncommitted.messages", 0L,
                         "ibm.mq.onqtime.short_period", -1L,
@@ -80,13 +78,14 @@ class QueueCollectionBuddyTest {
                 "DEV.QUEUE.1",
                 new HashMap<>(
                     ImmutableMap.of(
+                        "ibm.mq.expired.messages", 3L,
                         "ibm.mq.oldest.msg.age", -1L,
                         "ibm.mq.uncommitted.messages", 10L,
                         "ibm.mq.onqtime.short_period", -1L,
                         "ibm.mq.onqtime.long_period", -1L,
                         "ibm.mq.queue.depth", 1L))));
 
-    for (MetricData metric : otelTesting.getMetrics()) {
+    for (MetricData metric : producer.produce(Resource.empty())) {
       for (LongPointData d : metric.getLongGaugeData().getPoints()) {
         String queueName = d.getAttributes().get(MESSAGING_DESTINATION_NAME);
         Long expectedValue = expectedValues.get(queueName).remove(metric.getName());
@@ -103,7 +102,9 @@ class QueueCollectionBuddyTest {
   void testProcessPcfRequestAndPublishQMetricsForInquireQCmd() throws Exception {
     PCFMessage request = createPCFRequestForInquireQCmd();
     when(pcfMessageAgent.send(request)).thenReturn(createPCFResponseForInquireQCmd());
-    classUnderTest = new QueueCollectionBuddy(meter, new QueueCollectorSharedState());
+    MetricProducer producer =
+        new MetricProducer(Resource.empty(), InstrumentationScopeInfo.empty());
+    classUnderTest = new QueueCollectionBuddy(producer, new QueueCollectorSharedState());
     classUnderTest.processPcfRequestAndPublishQMetrics(
         collectorContext, request, "*", InquireQCmdCollector.ATTRIBUTES);
 
@@ -125,7 +126,7 @@ class QueueCollectionBuddyTest {
                         "ibm.mq.open.input.count", 3L,
                         "ibm.mq.open.output.count", 3L))));
 
-    for (MetricData metric : otelTesting.getMetrics()) {
+    for (MetricData metric : producer.produce(Resource.empty())) {
       for (LongPointData d : metric.getLongGaugeData().getPoints()) {
         String queueName = d.getAttributes().get(MESSAGING_DESTINATION_NAME);
         Long expectedValue = expectedValues.get(queueName).remove(metric.getName());
@@ -146,11 +147,13 @@ class QueueCollectionBuddyTest {
     sharedState.putQueueType("DEV.QUEUE.1", "local-transmission");
     PCFMessage request = createPCFRequestForResetQStatsCmd();
     when(pcfMessageAgent.send(request)).thenReturn(createPCFResponseForResetQStatsCmd());
-    classUnderTest = new QueueCollectionBuddy(meter, sharedState);
+    MetricProducer producer =
+        new MetricProducer(Resource.empty(), InstrumentationScopeInfo.empty());
+    classUnderTest = new QueueCollectionBuddy(producer, sharedState);
     classUnderTest.processPcfRequestAndPublishQMetrics(
         collectorContext, request, "*", ResetQStatsCmdCollector.ATTRIBUTES);
 
-    for (MetricData metric : otelTesting.getMetrics()) {
+    for (MetricData metric : producer.produce(Resource.empty())) {
       Iterator<LongPointData> iterator = metric.getLongGaugeData().getPoints().iterator();
       if (metric.getName().equals("ibm.mq.high.queue.depth")) {
         assertThat(iterator.next().getValue()).isEqualTo(10);
@@ -166,41 +169,44 @@ class QueueCollectionBuddyTest {
       PCFMessage:
       MQCFH [type: 1, strucLength: 36, version: 1, command: 41 (MQCMD_INQUIRE_Q_STATUS), msgSeqNumber: 1, control: 1, compCode: 0, reason: 0, parameterCount: 2]
       MQCFST [type: 4, strucLength: 24, parameter: 2016 (MQCA_Q_NAME), codedCharSetId: 0, stringLength: 1, string: *]
-      MQCFIL [type: 5, strucLength: 32, parameter: 1026 (MQIACF_Q_STATUS_ATTRS), count: 4, values: {2016, 1226, 1227, 1027}]
+      MQCFIL [type: 5, strucLength: 36, parameter: 1026 (MQIACF_Q_STATUS_ATTRS), count: 5, values: {2016, 1226, 1227, 1116, 1027}]
   */
   private static PCFMessage createPCFRequestForInquireQStatusCmd() {
     PCFMessage request = new PCFMessage(CMQCFC.MQCMD_INQUIRE_Q_STATUS);
     request.addParameter(CMQC.MQCA_Q_NAME, "*");
-    request.addParameter(CMQCFC.MQIACF_Q_STATUS_ATTRS, new int[] {2016, 1226, 1227, 1027});
+    request.addParameter(CMQCFC.MQIACF_Q_STATUS_ATTRS, new int[] {2016, 1226, 1227, 1116, 1027});
     return request;
   }
 
   /*
       0 = {PCFMessage@6026} "PCFMessage:
-      MQCFH [type: 2, strucLength: 36, version: 2, command: 41 (MQCMD_INQUIRE_Q_STATUS), msgSeqNumber: 1, control: 0, compCode: 0, reason: 0, parameterCount: 6]
+      MQCFH [type: 2, strucLength: 36, version: 2, command: 41 (MQCMD_INQUIRE_Q_STATUS), msgSeqNumber: 1, control: 0, compCode: 0, reason: 0, parameterCount: 7]
       MQCFST [type: 4, strucLength: 68, parameter: 2016 (MQCA_Q_NAME), codedCharSetId: 819, stringLength: 48, string: AMQ.5AF1608820C7D76E                            ]
       MQCFIN [type: 3, strucLength: 16, parameter: 1103 (MQIACF_Q_STATUS_TYPE), value: 1105]
       MQCFIN [type: 3, strucLength: 16, parameter: 3 (MQIA_CURRENT_Q_DEPTH), value: 12]
       MQCFIN [type: 3, strucLength: 16, parameter: 1227 (MQIACF_OLDEST_MSG_AGE), value: -1]
       MQCFIL [type: 5, strucLength: 24, parameter: 1226 (MQIACF_Q_TIME_INDICATOR), count: 2, values: {-1, -1}]
+      MQCFIN [type: 3, strucLength: 16, parameter: 1116 (MQIACF_EXPIRY_Q_COUNT), value: 0]
       MQCFIN [type: 3, strucLength: 16, parameter: 1027 (MQIACF_UNCOMMITTED_MSGS), value: 0]"
 
       1 = {PCFMessage@6029} "PCFMessage:
-      MQCFH [type: 2, strucLength: 36, version: 2, command: 41 (MQCMD_INQUIRE_Q_STATUS), msgSeqNumber: 2, control: 0, compCode: 0, reason: 0, parameterCount: 6]
+      MQCFH [type: 2, strucLength: 36, version: 2, command: 41 (MQCMD_INQUIRE_Q_STATUS), msgSeqNumber: 2, control: 0, compCode: 0, reason: 0, parameterCount: 7]
       MQCFST [type: 4, strucLength: 68, parameter: 2016 (MQCA_Q_NAME), codedCharSetId: 819, stringLength: 48, string: DEV.DEAD.LETTER.QUEUE                           ]
       MQCFIN [type: 3, strucLength: 16, parameter: 1103 (MQIACF_Q_STATUS_TYPE), value: 1105]
       MQCFIN [type: 3, strucLength: 16, parameter: 3 (MQIA_CURRENT_Q_DEPTH), value: 0]
       MQCFIN [type: 3, strucLength: 16, parameter: 1227 (MQIACF_OLDEST_MSG_AGE), value: -1]
       MQCFIL [type: 5, strucLength: 24, parameter: 1226 (MQIACF_Q_TIME_INDICATOR), count: 2, values: {-1, -1}]
+      MQCFIN [type: 3, strucLength: 16, parameter: 1116 (MQIACF_EXPIRY_Q_COUNT), value: 2]
       MQCFIN [type: 3, strucLength: 16, parameter: 1027 (MQIACF_UNCOMMITTED_MSGS), value: 0]"
 
       2 = {PCFMessage@6030} "PCFMessage:
-      MQCFH [type: 2, strucLength: 36, version: 2, command: 41 (MQCMD_INQUIRE_Q_STATUS), msgSeqNumber: 3, control: 0, compCode: 0, reason: 0, parameterCount: 6]
+      MQCFH [type: 2, strucLength: 36, version: 2, command: 41 (MQCMD_INQUIRE_Q_STATUS), msgSeqNumber: 3, control: 0, compCode: 0, reason: 0, parameterCount: 7]
       MQCFST [type: 4, strucLength: 68, parameter: 2016 (MQCA_Q_NAME), codedCharSetId: 819, stringLength: 48, string: DEV.QUEUE.1                                     ]
       MQCFIN [type: 3, strucLength: 16, parameter: 1103 (MQIACF_Q_STATUS_TYPE), value: 1105]
       MQCFIN [type: 3, strucLength: 16, parameter: 3 (MQIA_CURRENT_Q_DEPTH), value: 1]
       MQCFIN [type: 3, strucLength: 16, parameter: 1227 (MQIACF_OLDEST_MSG_AGE), value: -1]
       MQCFIL [type: 5, strucLength: 24, parameter: 1226 (MQIACF_Q_TIME_INDICATOR), count: 2, values: {-1, -1}]
+      MQCFIN [type: 3, strucLength: 16, parameter: 1116 (MQIACF_EXPIRY_Q_COUNT), value: 3]
       MQCFIN [type: 3, strucLength: 16, parameter: 1027 (MQIACF_UNCOMMITTED_MSGS), value: 0]"
   */
   private static PCFMessage[] createPCFResponseForInquireQStatusCmd() {
@@ -210,6 +216,7 @@ class QueueCollectionBuddyTest {
     response1.addParameter(CMQC.MQIA_CURRENT_Q_DEPTH, 12);
     response1.addParameter(CMQCFC.MQIACF_OLDEST_MSG_AGE, -1);
     response1.addParameter(CMQCFC.MQIACF_Q_TIME_INDICATOR, new int[] {-1, -1});
+    response1.addParameter(CMQCFC.MQIACF_EXPIRY_Q_COUNT, 0);
     response1.addParameter(CMQCFC.MQIACF_UNCOMMITTED_MSGS, 0);
 
     PCFMessage response2 = new PCFMessage(2, CMQCFC.MQCMD_INQUIRE_Q_STATUS, 2, false);
@@ -218,6 +225,7 @@ class QueueCollectionBuddyTest {
     response2.addParameter(CMQC.MQIA_CURRENT_Q_DEPTH, 0);
     response2.addParameter(CMQCFC.MQIACF_OLDEST_MSG_AGE, -1);
     response2.addParameter(CMQCFC.MQIACF_Q_TIME_INDICATOR, new int[] {-1, -1});
+    response2.addParameter(CMQCFC.MQIACF_EXPIRY_Q_COUNT, 2);
     response2.addParameter(CMQCFC.MQIACF_UNCOMMITTED_MSGS, 0);
 
     PCFMessage response3 = new PCFMessage(2, CMQCFC.MQCMD_INQUIRE_Q_STATUS, 1, false);
@@ -226,6 +234,7 @@ class QueueCollectionBuddyTest {
     response3.addParameter(CMQC.MQIA_CURRENT_Q_DEPTH, 1);
     response3.addParameter(CMQCFC.MQIACF_OLDEST_MSG_AGE, -1);
     response3.addParameter(CMQCFC.MQIACF_Q_TIME_INDICATOR, new int[] {-1, -1});
+    response3.addParameter(CMQCFC.MQIACF_EXPIRY_Q_COUNT, 3);
     response3.addParameter(CMQCFC.MQIACF_UNCOMMITTED_MSGS, 10);
 
     return new PCFMessage[] {response1, response2, response3};